requirements.txt
injector==0.22.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
bcrypt==4.3.0
click==8.2.1
fastapi[standard]
h11==0.16.0
httptools==0.6.4
httpx==0.28.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-multipart==0.0.20
python-slugify==8.0.4
PyYAML==6.0.2
SQLAlchemy==2.0.43
typing_extensions==4.14.1
uvicorn==0.35.0
passlib[bcrypt]
pyjwt==2.10.1
greenlet==3.2.3
environs==14.2.0
psycopg2-binary==2.9.10
pytz==2025.2
sqladmin==0.21.0
mypy==1.18.2
types-pytz==2025.2.0.20250809
